requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
python-jose==3.3.0
passlib==1.7.4
tinydb==4.8.0
boto3==1.26.137
python-dotenv==0.19.0
bcrypt==3.2.0
pydantic==1.8.2
transformers==4.30.2
torch==2.0.1
scikit-learn==1.0.2
pytesseract==0.3.10
pdf2image==1.16.3
spacy==3.5.3 